refactor: moved CustomVCSInput and CustomVCSUpdateInput to separate package vcs
