+++ spacelift/internal/structs/vcs/input.go
@@ -1,0 +1,12 @@
+package vcs
+
+import "github.com/shurcooL/graphql"
+
+// CustomVCSInput represents the custom VCS input data.
+type CustomVCSInput struct {
+	Name        graphql.String    `json:"name"`
+	SpaceID     graphql.ID        `json:"spaceID"`
+	Labels      *[]graphql.String `json:"labels"`
+	Description *graphql.String   `json:"description"`
+	IsDefault   *graphql.Boolean  `json:"isDefault"`
+}
