+++ spacelift/internal/structs/vcs/update.go
@@ -1,0 +1,11 @@
+package vcs
+
+import "github.com/shurcooL/graphql"
+
+// CustomVCSUpdateInput represents the custom VCS update input data.
+type CustomVCSUpdateInput struct {
+	ID          graphql.ID        `json:"id"`
+	SpaceID     graphql.ID        `json:"space"`
+	Labels      *[]graphql.String `json:"labels"`
+	Description *graphql.String   `json:"description"`
+}
